implemented JWT auth

+++ task.go
@@ -15,7 +15,11 @@ type TaskDb struct {
 	db *sql.DB
 }
 
-func (taskDb *TaskDb) AddTask(task Task) error {
+func (taskDb *TaskDb) AddTask(task Task, userName string) error {
+	if userName == "" {
+		return fmt.Errorf("UserName is empty")
+
+	}
 	if task.TaskId == "" {
 		return fmt.Errorf("TaskId is empty")
 
@@ -25,17 +29,23 @@ func (taskDb *TaskDb) AddTask(task Task) error {
 
 	}
 
-	_, err := taskDb.db.Exec("insert into tasks (taskId,taskName) values (?,?)", task.TaskId, task.TaskName)
+	_, err := taskDb.db.Exec("insert into tasks (taskId,taskName,owner) values (?,?,?)", task.TaskId, task.TaskName, userName)
 	if err != nil {
-		return err
+		fmt.Println(err)
+		return fmt.Errorf("Unable to add task")
 	}
 	return nil
 
 }
-func (taskDb *TaskDb) GetTasks() ([]Task, error) {
-	query, err := taskDb.db.Query("select taskId as TaskId,taskName as TaskName,CreatedDate as CreatedDate from tasks")
+func (taskDb *TaskDb) GetTasks(userName string) ([]Task, error) {
+	if userName == "" {
+		return nil, fmt.Errorf("UserName is empty")
+
+	}
+	query, err := taskDb.db.Query("select taskId as TaskId,taskName as TaskName,CreatedDate as CreatedDate from tasks where owner=?", userName)
 	if err != nil {
-		return nil, err
+		fmt.Println(err)
+		return nil, fmt.Errorf("Unable to fetch tasks")
 	}
 	defer query.Close()
 	var tasks []Task
@@ -48,7 +58,7 @@ func (taskDb *TaskDb) GetTasks() ([]Task, error) {
 	}
 	return tasks, nil
 }
-func (taskDb *TaskDb) EditTask(task Task) error {
+func (taskDb *TaskDb) EditTask(task Task, userName string) error {
 	if task.TaskId == "" {
 		return fmt.Errorf("TaskId is empty")
 
@@ -57,21 +67,31 @@ func (taskDb *TaskDb) EditTask(task Task) error {
 		return fmt.Errorf("TaskName is empty")
 
 	}
-	_, err := taskDb.db.Exec("update tasks set taskName=? where taskId=?", task.TaskName, task.TaskId)
+	if userName == "" {
+		return fmt.Errorf("UserName is empty")
+
+	}
+	_, err := taskDb.db.Exec("update tasks set taskName=? where taskId=? and owner=?", task.TaskName, task.TaskId, userName)
 	if err != nil {
-		return err
+		fmt.Println(err)
+		return fmt.Errorf("Unable to update task with taskId %s", task.TaskId)
 	}
 	return nil
 
 }
-func (taskDb *TaskDb) DeleteTask(taskId string) error {
+func (taskDb *TaskDb) DeleteTask(taskId string, userName string) error {
 	if taskId == "" {
 		return fmt.Errorf("TaskId is empty")
 
 	}
-	res, err := taskDb.db.Exec("delete from tasks where taskId=?", taskId)
+	if userName == "" {
+		return fmt.Errorf("UserName is empty")
+
+	}
+	res, err := taskDb.db.Exec("delete from tasks where taskId=? and owner=?", taskId, userName)
 	if err != nil {
-		return err
+		fmt.Println(err)
+		return fmt.Errorf("Unable to delete task with taskId %s", taskId)
 	}
 	rowsAffected, err := res.RowsAffected()
 	if err != nil {
